Dev backend

+++ BackEnd/src/main/java/com/dokb/DoKB/history/domain/HistoryApi.java
@@ -1,14 +1,10 @@
 package com.dokb.DoKB.history.domain;
 
-import com.dokb.DoKB.account.domain.Account;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -17,18 +13,18 @@ import java.time.LocalDateTime;
 @Builder
 public class HistoryApi {
 
-    private Long id;
+	private Long id;
 
-    private String opponentAccount;
+	private String opponentAccount;
 
-    private LocalDateTime dealDate;
+	private LocalDateTime dealDate;
 
-    private String inOut;
+	private String inOut;
 
-    private BigDecimal amount;
+	private Long amount;
 
-    private BigDecimal balance;
+	private Long balance;
 
-    private String accountNumber;
+	private String accountNumber;
 
 }
